Move top-balances test into separate file
